+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add tooltip to openFullscreen button ([#31](https://github.com/scm-manager/scm-statistic-plugin/pull/31))
+
 ### Fixed
 - Add tab stops to statistics page zoom action ([#29](https://github.com/scm-manager/scm-statistic-plugin/pull/29))